Refactor/restructure form components
